Fix (unreported) wrong definition of `ssize_t` for windows.
Since at least MSVC2010 there is a `SSIZE_T` available for windows, use
it to typedef `ssize_t` on this platform.

+++ source/blender/blenlib/BLI_winstuff.h
@@ -84,7 +84,7 @@ typedef unsigned int mode_t;
 /* python uses HAVE_SSIZE_T */
 #  ifndef HAVE_SSIZE_T
 #    define HAVE_SSIZE_T 1
-typedef long ssize_t;
+typedef SSIZE_T ssize_t;
 #  endif
 #endif
